more compact rsp-api-doc format with :id
I can justify this step saying that it's fine that we have to specs but
wouldn't it be great if this also served as documentation? Enter rspec
api doc. First make a commit WITHOUT grouping by /:id but only by
generating and showing the docs. Then add this nice thing