Assign random value to attributes
Now the only thing left hard-coded are these instances. What is bad about
this? Two things: we are writing tests for VERY SPECIFIC cases in the
database, and also our documentation will ALWAYS show these examples, as
though these were the only things ever returned by our API. But actually
we can express this in a more meaningful way. All we need is to create
some instances where the attributes have any random value for their type.
And we already know the types, they are in our attributes list! And not
only that, but we also know which attributes can be nil, so for those
we don't even have to always set a value (year). So we can write the
following (note that for now, one instance is enough for all tests)

Now, run rspec and look at the log and you will see random values!