Complete rewrite, removed rspec-api-documentation
RSpec API has been rewritten from scratch, implementing the whole
DSL from `resource` to `request` and `respond_to`, in a way that is
compatible with the previous DSL but does not depend on rspec-api-
documentation.

This makes RSpec API more flexible in terms of what goes inside the
`it` blocks and how fixtures are created. Also, for the first time,
it supports not just local API (built with ActiveRecord and accessed
with RackTest), but remote API (accessed with Faraday)